pyproject.toml
[project]
name = "hvostiki-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "celery (>=5.6.1,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]

[project.optional-dependencies]
dev = [
    "pytest (>=8.0)",
    "pytest-django (>=4.8)",
    "django-debug-toolbar (>=4.4)",
    "black (>=24.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
